driver: use simplesurance/pq instead of lib/pq
The migrate library is importing a specific postgres driver, lib/pq.
It's only possible to register 1 driver for an implementation in the
database/sql stdlib.

This causes the problem that users of the library can not choose their
own implementation of the postgres driver when they are using migrate.

To be able to use simplesurance/pq instead of lib/pq in our application,
the import in the migrate lib must be changed.